General work

+++ remote/clean.sh
@@ -1,5 +1,26 @@
 #!/bin/bash
-docker rm $(docker ps -a -q)
+#docker rm $(docker ps -a -q)
+
+
+if [  -f /tmp/.install-in-progress ]
+then
+    echo "Install in progress."
+    exit 0
+fi
+
+if [ -f /tmp/.initializing-in-progress ]
+then
+    echo "Still initialzing component"
+    exit 0
+fi
+
+
+if test $(find "/tmp/.restart-in-progress" -mmin -30)
+then
+    echo "Restart in progress"
+    exit 0
+fi
+
 docker rmi $(docker images -a | grep "^<none>" | awk '{print $3}')
 
 if [ -f /home/easydeploy/deployment/clean.sh ]
